g_cwd に関する関数を g_cwd.c に移動
