+++ Makefile
@@ -10,7 +10,7 @@ HEADER_FILES = minishell.h
 SRCS = cmd_cmd_invocation.c cmd_cmd_invocation2.c cmd_exec_command.c		\
 	cmd_exec_commands.c cmd_pipe.c cmd_redirection.c convert_ast2cmdinvo.c	\
 	env.c exec.c lexer1.c lexer2.c minishell.c parse1.c parse2.c			\
-	parse_utils.c parse_utils2.c path.c string_node2string.c				\
+	parse_utils.c parse_utils2.c path.c g_cwd.c string_node2string.c		\
 	expand_env_var.c expand_string_node.c split_expanded_str.c				\
 	cmd_status.c signal.c env_setter.c minishell_error_msg.c				\
 	builtin.c builtin_cd.c builtin_echo.c builtin_env.c builtin_exit.c		\
